Refactor to remove reducer layer for roles
