+++ src/actions/Types.js
@@ -79,6 +79,7 @@ export const REMOVE_ADJACENT_SITE = "REMOVE_ADJACENT_SITE";
 export const SET_STOP_PLACE_LOADING = "SET_STOP_PLACE_LOADING";
 
 // UserActions
+export const UPDATED_AUTH = "UPDATED_AUTH";
 export const NAVIGATE_TO = "NAVIGATE_TO";
 export const TOGGLED_IS_CREATING_NEW_STOP = "TOGGLED_IS_CREATING_NEW_STOP";
 export const APPLIED_STOPTYPE_SEARCH_FILTER = "APPLIED_STOPTYPE_SEARCH_FILTER";
@@ -198,8 +199,3 @@ export const SETUP_NEW_GROUP = "SETUP_NEW_GROUP";
 export const CREATED_NEW_GROUP_OF_STOP_PLACES =
   "CREATED_NEW_GROUP_OF_STOP_PLACES";
 export const ERROR_NEW_GROUP = "ERROR_NEW_GROUP";
-
-// RolesActions
-export const UPDATED_AUTH = "UPDATED_AUTH";
-export const UPDATED_ALLOW_NEW_STOPS_EVERYWHERE =
-  "UPDATED_ALLOW_NEW_STOPS_EVERYWHERE";
