+++ src/components/EditParentStopPage/EditParentGeneral.js
@@ -42,8 +42,8 @@ import {
   saveParentStopPlace,
   terminateStop,
 } from "../../actions/TiamatActions";
-import Routes from "../../routes";
 import SettingsManager from "../../singletons/SettingsManager";
+import { getStopPermissions } from "../../utils/permissionsUtils";
 
 class EditParentGeneral extends React.Component {
   constructor(props) {
@@ -513,7 +513,7 @@ class EditParentGeneral extends React.Component {
   }
 }
 
-const mapStateToProps = ({ stopPlace, mapUtils, roles, user }) => ({
+const mapStateToProps = ({ stopPlace, mapUtils, user }) => ({
   stopPlace: stopPlace.current,
   versions: stopPlace.versions,
   activeMap: mapUtils.activeMap,
@@ -521,7 +521,7 @@ const mapStateToProps = ({ stopPlace, mapUtils, roles, user }) => ({
   removeStopPlaceFromParentOpen: mapUtils.removeStopPlaceFromParentOpen,
   removingStopPlaceFromParentId: mapUtils.removingStopPlaceFromParentId,
   adjacentStopDialogOpen: user.adjacentStopDialogOpen,
-  canDeleteStop: getIn(roles, ["allowanceInfo", "canDeleteStop"], false),
+  canDeleteStop: getStopPermissions(stopPlace).canDelete,
   deleteStopDialogOpen: mapUtils.deleteStopDialogOpen,
   originalStopPlace: stopPlace.originalCurrent,
   serverTimeDiff: user.serverTimeDiff,
