+++ src/components/GroupOfStopPlaces/EditGroupOfStopPlaces.js
@@ -31,6 +31,7 @@ import Routes from "../../routes/";
 import { UserActions, StopPlacesGroupActions } from "../../actions/";
 import ConfirmDialog from "../Dialogs/ConfirmDialog";
 import { getIn } from "../../utils/";
+import { getStopPermissions } from "../../utils/permissionsUtils";
 
 class EditGroupOfStopPlaces extends Component {
   constructor(props) {
@@ -284,12 +285,12 @@ class EditGroupOfStopPlaces extends Component {
   }
 }
 
-const mapStateToProps = ({ stopPlacesGroup, roles }) => ({
+const mapStateToProps = ({ stopPlacesGroup }) => ({
   isModified: stopPlacesGroup.isModified,
   groupOfStopPlaces: stopPlacesGroup.current,
   originalGOS: stopPlacesGroup.original,
-  canEdit: getIn(roles, ["allowanceInfo", "canEdit"], false),
-  canDelete: getIn(roles, ["allowanceInfo", "canDeleteStop"], false),
+  canEdit: getStopPermissions(stopPlacesGroup.current).canEdit,
+  canDelete: getStopPermissions(stopPlacesGroup.current).canDelete,
 });
 
 export default connect(mapStateToProps)(injectIntl(EditGroupOfStopPlaces));
